requirements.txt
aiogram==2.23.1
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
Babel==2.9.1
certifi==2022.9.24
charset-normalizer==2.1.1
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.2
pytz==2022.6
typing_extensions==4.4.0
yarl==1.8.1
aiogram==2.23.1
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
Babel==2.9.1
certifi==2022.9.24
charset-normalizer==2.1.1
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.2
pytz==2022.6
typing_extensions==4.4.0
yarl==1.8.1
aiogram==2.23.1
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
Babel==2.9.1
certifi==2022.9.24
charset-normalizer==2.1.1
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.2
pytz==2022.6
typing_extensions==4.4.0
yarl==1.8.1
aiogram==2.23.1
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
Babel==2.9.1
backcall==0.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
cycler==0.11.0
decorator==5.1.1
fonttools==4.38.0
frozenlist==1.3.3
idna==3.4
ipython==7.34.0
jedi==0.18.1
kiwisolver==1.4.4
magic-filter==1.0.9
matplotlib==3.5.3
matplotlib-inline==0.1.6
multidict==6.0.2
numpy==1.21.6
opencv-python==4.6.0.66
packaging==21.3
pandas==1.3.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
prompt-toolkit==3.0.32
psutil==5.9.4
ptyprocess==0.7.0
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
requests==2.28.1
scipy==1.7.3
seaborn==0.12.1
six==1.16.0
torch==1.13.0+cpu
torchaudio==0.13.0+cpu
torchvision==0.14.0+cpu
tqdm==4.64.1
traitlets==5.5.0
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
yarl==1.8.1
aiogram==2.23.1
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
Babel==2.9.1
backcall==0.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
cycler==0.11.0
decorator==5.1.1
fonttools==4.38.0
frozenlist==1.3.3
idna==3.4
ipython==7.34.0
jedi==0.18.1
kiwisolver==1.4.4
magic-filter==1.0.9
matplotlib==3.5.3
matplotlib-inline==0.1.6
multidict==6.0.2
numpy==1.21.6
opencv-python==4.6.0.66
packaging==21.3
pandas==1.3.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
prompt-toolkit==3.0.32
psutil==5.9.4
ptyprocess==0.7.0
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
requests==2.28.1
scipy==1.7.3
seaborn==0.12.1
six==1.16.0
torch==1.13.0+cpu
torchaudio==0.13.0+cpu
torchvision==0.14.0+cpu
tqdm==4.64.1
traitlets==5.5.0
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
yarl==1.8.1
aiogram==2.23.1
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
Babel==2.9.1
backcall==0.2.0
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cycler==0.11.0
decorator==5.1.1
fonttools==4.38.0
frozenlist==1.3.3
idna==3.4
importlib-metadata==5.0.0
ipython==7.34.0
isort==5.10.1
jedi==0.18.1
kiwisolver==1.4.4
magic-filter==1.0.9
matplotlib==3.5.3
matplotlib-inline==0.1.6
multidict==6.0.2
mypy-extensions==0.4.3
numpy==1.21.6
opencv-python==4.6.0.66
packaging==21.3
pandas==1.3.5
parso==0.8.3
pathspec==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.4
prompt-toolkit==3.0.32
psutil==5.9.4
ptyprocess==0.7.0
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
requests==2.28.1
scipy==1.7.3
seaborn==0.12.1
six==1.16.0
tomli==2.0.1
torch==1.13.0+cpu
torchaudio==0.13.0+cpu
torchvision==0.14.0+cpu
tqdm==4.64.1
traitlets==5.5.0
typed-ast==1.5.4
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
yarl==1.8.1
zipp==3.10.0
